SDC-4990: Invalid IP Address Stops Pipeline Even When Geo IP Lookup Configured as "Send to Error"
The issue was that we are catching all exceptions to log and react to
them and then propagating only subset of classes. So it happened that we
forgot to properly propagate OnRecordErrorException that actually causes
records to go to error stream rather then stopping a pipeline.

Change-Id: Ida2fd3987b21aec7a8a14162e42b17986c29ac8a
Reviewed-on: https://review.streamsets.net/6449
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
(cherry picked from commit 45680268c1d2634214bca4e8f3ea9536fc76edb2)
Reviewed-on: https://review.streamsets.net/6499
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>