SDC-5232: HADOOPFS_59 - Recovery failed to rename old _tmp_ files
The reason why the regexp replacement did not worked again is that the
user expression had additional unexpected characters like '_'. What
about explicitly iterating what characters are allowed, we simply say
which are not? E.g. only } is really not allowed inside.

That will still not work for cases when someone do something like
${str:replace('}')}, but it seems fairly unlikely that someone would put
the '}' directly into the expression and in such case the user can still
disable the recovery altogether.

Change-Id: Id395bf7706bad2ca18dec383c40d789018a67722
Reviewed-on: https://review.streamsets.net/6450
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>