SDC-5237. Support for mapping different subjects In Pipeline Acls.
This patch adds a new APi in ACl Store to update the subject names to different names. This will be really useful when SSO is enabled and subjects (users/groups) are completely changed. A Rest API to do this is also added /rest/v1/pipelines/updateSubjectsInAcls. The logic will sequentially go through all pipelines and visit acl information and update the subject with new mapped subject in the following information:

1. Resource Owner
2. Last Modified By
3. Permissions -> SubjectId, Last Modified By

Change-Id: If70863232da0284a8d4a2ef5223b52b629996aa7
Reviewed-on: https://review.streamsets.net/6458
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>
Reviewed-by: Harikiran Nayak <hari@streamsets.com>