SDC-2063: Processor for converting MAP record to LIST_MAP/LIST record
The change looks bigger then it really is as there is a lot of
configuration boiler plate code. The main functionality is "just" a few
lines in the FieldOrderProcessor.

The processor works as follows - user need to specify order in which the
keys should be ordered and then we will create ordered LIST_MAP in that
order.

There are two actions that user might select - in case that some fields
are missing (e.g. order should be /a and then /b, but /b is missing in
source record) with possibility to send the record to either error
stream or alternatively use a default value that needs to be specified.
Second action is for the case when record have an extra fields (e.g.
sort order is /a, /b but the record have /a, /b and /c). Actions are
very similar - either send record to error or alternatively ignore the
extra fields. The processor will never put the extra fields to the
output record as we would not be able to guarantee their order.

Change-Id: Iaac97958b1fdc7d56b47d956d2dcc81347fe2ea7
Reviewed-on: https://review.streamsets.net/6454
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>