SDC-5301. Multi Table Jdbc: Record Header 'jdbc.tables' is not populated for SQL Server.
Currently We rely on ResultSetMetadata#getTables to get table name for the current record. It seems like MSSQL Server by default returns empty when we call that API. In Multi Table Jdbc Consumer given that we already know what table we are already reading rows so use that information instead of relying on ResultSetMetadata#getTables API. This patch adds this logic to MultiTableConsumer, but does not touch this functionality for Jdbc Query Consumer as the record is generated from the result of a query which could join multiple tables.

Change-Id: I34fa721219d06453968ea6a7f3e56906785da8a8
Reviewed-on: https://review.streamsets.net/6555
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
(cherry picked from commit 995fa044fb41b80dc55763d50efaa60f1faef89b)
Reviewed-on: https://review.streamsets.net/6564
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>