SDC-5248: Provide mechanism for a stage to declare whether compatible JDKs
This patch adds a new optional property to already existing file
data-collector-library.properties that allows stage to express JVM
version regular expression for which the whole stage library is not
compatible with. The JDK8-only stage libraries were removed from the
blacklist and converted to use this property instead.

If the user is on JDK8 then everything is seamless for them - installing
the stage library will make it immediately available without a need to
manually enable it. The behavior for JDK7 remains the same as if the
stage would be blacklisted in sdc.properties.

I've tested the patch both on JDK7 and JKD8 and it works as expected.
I'm not sure how easily automate that though.

Change-Id: Ic86c75f0f62f0719d1e78345ddaf544bcc764bf6
Reviewed-on: https://review.streamsets.net/6560
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
(cherry picked from commit 2fa45eccd63b0ce3310c9ddd4200254c0724c9dd)
Reviewed-on: https://review.streamsets.net/6563
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>