SDC-4928: Share class loaders for the same stages across all runners
Rather then creating completely new stages with new stage definitions,
this patch updates to logic to rather duplicate the first runner.
Duplicate runners have their own instance of the stage - and hence the
stages still doesn't have to be thread safe - but will share the same
stage definition and hence the same class loader. This includes stages
with private class loaders as well.

On the release side, this patch changes the release order so that
runners with id >0 are release first. They release doesn't release the
class loader as they do not own the class loader. Only when releasing
the runner with id=0 we release the shared class loaders.

Change-Id: Ic79cf6e0b3ab279d83ed31710e369dc29ddbbdfd
Reviewed-on: https://review.streamsets.net/6701
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>