SDC-5368: Runner id is not always correctly printed in the logs
The issue was that we were printing runner id from the first pipe which
happens to be a ObserverPipe that doesn't have a notion of runner id. We
have to get pipe with index 3 that is a StagePipe and hence have the
notion of runner.

Perhaps a better solution would be to create a custom class for the
runner that would wrap the List<Pipe>, but I didn't want to do that as
part of this patch. I'll probably wait on another reason to promote this
concept to a standalone class and then refactor this.

Change-Id: I5fb34d98036077c793f18ce48913c7800a5580ad
Reviewed-on: https://review.streamsets.net/6706
Reviewed-by: Jeff Evans <jeff@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>