SDC-5402.  Expose Record ELs for MapR JSON Destination table name.
Added support for ELs in the tableName, which permits the customer to select a
field or attribute which will provide the name of the table for this record.

For example, in the Multi-Table JDBC case this expression:
${record:attribute(‘jdbc.tables’)}
will select the table name from each record - and the
MapR JSON Destination will send that record to the selected table.

If the table does not exist, and the “Create Table” option is selected in
the UI, the table will be created.

In addition, the ELs can have prefixes or suffixes (or both) such as:
foo_${record:attribute(‘jdbc.tables’)}_bar

Also, field values are permitted for example - ${record:value('/store_num’)}

Currently only one EL is permitted in the table name field
but combinations of record and string expressions are permitted
within one EL expression.  for example:
${str:toUpper(record:attribute('jdbc.tables’))}_NEW
forces the table name to upper case, and appends the suffix _NEW.

if the EL resolves to an empty String a StageException occurs.
if the EL cannot be parsed a StageException occurs.

Change-Id: I89ae23d4fbc2edfbd844af2945b9310e08fe3b0c
Reviewed-on: https://review.streamsets.net/6877
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>