SDC-5481: Add histogram with available runners to pipeline metrics
The idea is to help the user understand how well are the runners used.
After having several conversations with Alejandro Abdelnur, we've ended
up on providing Histogram showing number of available runners
(work-in-progress screenshot attached). If the number is close to
configured number of runners, then the user knows that most of the
runners sit idle most of the time. If on the other hand, the number is
close to zero, then the user might consider increasing number of
runners.

Change-Id: I3f3329e257a9f62ec3b63d929bdb6050a1019f83
Reviewed-on: https://review.streamsets.net/6963
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>