SDC-5499 — createDateFromStringTZ
CreateDateFromStringTZ() takes 3 parameters and creates a Date
object with the timezone applied to the provided datetime.
The String passed in is assumed to have been created in the
provided timezone (second parameter), and is then converted
to a Date object.  Hence, “2017-01-01 11:00:05” on a machine
running with the JVM in Pacific time can be changed to Eastern
time, and still be the same value.  Basically, this provides
timezone “override”.

First Parameter is the String representation of the datetime.
This representation must match the format (3rd parameter)
and must not have a timezone in the datetime String.

The second parameter is the timezone to be applied to this
datetime.  This time zone should not be a 3-letter timezone
(except for GMT/UTC).  A list of some appropriate timezones
can be found here:

https://www.vmware.com/support/developer/vc-sdk/visdk400pubs/ReferenceGuide/timezone.html

The third parameter is the date format which describes the
datetime provided in the first parameter.   This date format
must not have Z, z, or x specified for the timezone.

The function returns a Date object.

Note that using toString on the Date object will display
the date in the JVM’s local time. Internally, the Date
object maintains the datetime in UTC.

Change-Id: I483776f97714a9c6a2f0be6106dd85b510aead74
Reviewed-on: https://review.streamsets.net/7027
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
(cherry picked from commit 39d8e20af7ae4c96d9601aef62c373a68468ddbb)
Reviewed-on: https://review.streamsets.net/7083