SDC-5457: Missing ${PATTERN} from file tail path causes StringIndexOutOfBoundsException
We were already properly checking that the ${PATTERN} is present and
issuing validation error if it was not. Sadly the way our validations
were written, not having the token did not stop some further validations
from happening and thus we run in the exception.

I've refactored alltoken validations to be done in a single method and
that further validations won't be run if the pre-requisite check isn't
satisfied.

Change-Id: I51918b561eca248e07e92920dcfee95aa45abdc1
Reviewed-on: https://review.streamsets.net/7069
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Bob Plotts <bob@streamsets.com>