SDC-5581. NPE in Salesforce Origin
NPE is caused by use of SystemModStamp rather than SystemModstamp as
the offset field - note the case difference on the S in the middle.
Salesforce field names are case-insensitive, but we search for a value
in the returned XML and that search is case sensitive.

Fix is to code a case-insensitive search and also check the returned
element in case the offset field is present in the returned fields
(meaning it was not present in the query).

Change-Id: I2cc8d7a4c7fe121652976c6eacac9626818ffb8f
Reviewed-on: https://review.streamsets.net/7152
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>