SDC-5383: Create first class concept for Runner
This is a bit bigger change then anticipated. The pipe list was
converted into standalone class called PipeRunner that exposes forEach
and forEachNoException methods to operate on the pipes. Benefit of this
approach is that we can easily track runner id and we can set it in MDC
before each operation which simplifies it's use.

Change-Id: Ic5057f6a11b3e4328376401b01308d7350cdb59e
Reviewed-on: https://review.streamsets.net/7162
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>