SDC-5601. Multi Table Jdbc : Use Executor Completion Service to detect worker threads status.
This patch used ExecutorCompletionService to detect and handle worker thread completion status. Before this there was a while loop going through each worker future to see if it is done, this patch gets rid of the while loop and simply uses executorCompletionService.poll() to detect the future which completes first.

Change-Id: I6e6c7849754fbd81960fa1bfd6f91c30c5c39558
Reviewed-on: https://review.streamsets.net/7187
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>