SDC-5527: setup-mapr for 5.2 error
We've switched blacklist property from a single line to a multi-line a
while back to make it easier to spot what is enabled and what is not as
we had increasing amount of JDK-8 stages. Since we've removed the JDK-8
stages from list a while ago, there is no longer a reason to keep the
property as a multi-line.

This patch removes all the advanced bash scripts that were trying to
work properly with the multi-line and rather keeps everything on a
single line thus removing the need to solve the edge condition that was
failing.

Change-Id: I6e1f935fb61fe466caf7d7f15eb0b62d2fe0b464
Reviewed-on: https://review.streamsets.net/7275
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>