SDC-5664. SDCClassLoader does not pass ordered URLs to URLClassLoader.
This patches fixes the case where the urls passed to base URLClassLoader from SDCClassLoader is not ordered (i.e stagelib/protolib at the front of the list and other third party/stage related libs ordered after that). This will help deterministic ordering on the jars we load and also allow us to override third party classes in case we wish to provide specific implementations.

Change-Id: I8db4a833ebe6ffdaa0a17cc31c6256d9b695d268
Reviewed-on: https://review.streamsets.net/7303
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Harikiran Nayak <hari@streamsets.com>