SDC-5560: SDC install on a MapR Client host
From some reason MapR's client and cluster package differs in content of
the /opt/mapr/lib directory that we're simplinking to our stage
libraries. Some of the jars that we require are not there in the
mapr-client package.

This patch solves the problem by explicitly symlinking all jars from
hadoop common space which at least on my test environment got the mapr
fs destination working.

Change-Id: Iacff2b06ae744c55d37ff494ca53b44a7436a059
Reviewed-on: https://review.streamsets.net/7309
Reviewed-by: Rupal Shah <rupal@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>