SDC-5410: MapReduce Executor not working with MapR
I've noticed two different issues that could lead to this or similar
issues:

1) MapR client package does not ship mapreduce libraries in
/opt/mapr/lib, so we need to import them from Hadoop instalation that is
next to this directory.

2) MapR does ship contradicting jar files in /opt/mapr/lib - e.g. one
can't take the whole directory and put it into the same classpath as
based on the jar order, the behavior will be different. In our case, we
don't support the old jobtracker APIs, but that is what can be picked up
automatically since the jars sits on the classpath.

Change-Id: I8938ecc2d270985bab135a3c6c3355ddff57ad87
Reviewed-on: https://review.streamsets.net/7312
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>