SDC-5696. Option to store all fields in paginated http source
In the current and default mode, records are parsed only from the result
field path list field. Enabling this option will also copy all *other*
fields to the resulting records. This is similar to pivoting on the
result field path with the Keep Existing Fields option in the Field
Pivot Processor.

When enabled, the result field path will remain at its original position
in the output record, rather than becoming the new root of the record.

Change-Id: I214dffb4cd6b079a281833da0b4a03f12ccdc595
Reviewed-on: https://review.streamsets.net/7401
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>