SDC-5708.  Pipeline Finisher Executor.
Created new Executor to receive an event.
When an event is received - the pipeline will
transition to "FINISHED" state.

The Executor does not check the event contents.
The user must ensure that the correct event is
passed to the executor.  The best way to do this
is by putting a Stream Selector in the event stream
before the Pipeline Finisher Executor.  Use this EL
to select the no-more-data events.

${str:contains(record:attribute("sdc.event.type"), "no-more-data")}

Change-Id: I2259ca2dfc36b233467b97d91a6675d1fd8396a1
Reviewed-on: https://review.streamsets.net/7412
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>