SDC-5726: Pipeline name and user not visible in the logs from SDC installed with CM
Added one last override that is needed to make the logs show in our UI
when used on CM managed cluster.

I've verified that both all and pipeline logs works properly. This
change will however break CM itself - e.g. CM will not be able to parse
SDC logs any more.

I've tried to look into it a bit more, but unless Cloudera resolves
OPSAPS-21675 that adds support for custom log patters in CSD, it seems
that where the logs will work is a binary choice. It will either be CM
or SDC.

Given the extended facilities that we provide in logs, I think that
breaking the CM view and making the logs available directly in SDC
have more value at this point.

The "...appender.streamsets.layout...." are properties needed for
our backend handling of the log parsing and "..appender.RFA.layout..."
is need for CM as it forces this appender name.

Change-Id: I97af6ee71a21889518e41d8b356a2cef64ceb444
Reviewed-on: https://review.streamsets.net/7437
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>