SDC-5737: New CSD is not able to start older parcels
This is because of SDC-5527 where we made simplification changes -
instead of having blacklist on multiple lines which is a bit hard to
deal with from bash, we switched to exactly one line. That also
simplified the logic in the control.sh script.

Sadly the CSD is needs to work with more then one version of parcel -
this is important for example during upgrades where people need to
upgrade CSD first and then parcel.

This patch is a revert of the SDC-5527's control.sh change. We will be
able to get back to the simpler code once we stop supporting SDC <2.5.

Change-Id: Ic86c8759e8f42d1f114ec81764ba6cf620b97780
Reviewed-on: https://review.streamsets.net/7438
Reviewed-by: Bob Plotts <bob@streamsets.com>