SDC-5750: StandaloneAndClusterPipelineManager propagates incorrect user when starting pipeline
When starting a pipeline the manager was propagating the user who have
created the pipeline rather then the user who started the pipeline.
We're already storing the user who started the pipeline in the history
file, but since we're not displaying that in the UI yet, nobody noticed
that until now. This will however break the logic that we introduced for
SDC-3704 and hence we need to fix that.

There are two events that causes pipeline to start - first is direct
REST call from user which this patch resolves. The second one is a
pipeline auto-start on a SDC start up. This code already looks into the
history file to see what is the user that actually started the pipeline
to start it with the same user. Hence the second portion doesn't need to
be fixed.

Change-Id: If3da7fb3c5ea6364c114f5a12e511fb75e4bb15c
Reviewed-on: https://review.streamsets.net/7456
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>