SDC-5751. Update name field in Pipeline Configuration, Pipeline Info & Pipeline State bean
To support pipeline rename and multi tenant (multiple users can create
private pipeline with same name) and to reduce impact on backward
compatibility we autogenerated unique values to name property and added
"title" for user provided display label for pipeline.

And UI is updated to show "name" property as 'ID'. This is causing
confusion to the user since in UI it is displayed as 'ID' but in
pipeline configuration/info/state json it is 'name' property.

To reduce the confusion I have updated 'name' property to 'pipelineId'
in pipeline configuration/info/state beans and deprecated 'name'
property in their respective json beans.

Change-Id: Ideefc6721ea378d846215bcd4cf018c5796b4eab
Reviewed-on: https://review.streamsets.net/7434
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Reviewed-by: Harikiran Nayak <hari@streamsets.com>