SDC-5763: Dev Data Generator - Allow configuring event-specific details
I've added ability to specify name for the event - it was previously
compile time constant, now it's exposed to user for configuration.

The fields of the event are already configurable as we use the existing
configuration for fields (e.g. we send the same record to both output
and event streams).

Change-Id: Iabe41ea58b35f50247243d8c2877318a4200c4b1
Reviewed-on: https://review.streamsets.net/7495
Reviewed-by: Loretta Chen <loretta@streamsets.com>