SDC-5773: Runner should have user for each operation rather then use "global" user
Up until now the Runner class had a concept of owner (a user) that was
assigned at the time of creation and was never changed. This however had
several shortcomings:

* Runners are reused so it can easily happen that when switching users,
  the pipeline is executed under incorrect user.
* Different operations of the same runner can be executed by a different
  user - e.g. runner can be started by 'jarcec', but stopped by
  'arvind'.
* Since the user is never changed, the pipeline's history file
  contains only a single user name for all operations, regardless of who
  actually executed them.

This patch changes that behavior - instead of the runner having a global
concept of owner, each operation now contains the user who executed it.
This will also solve several independent problems we've seen while
investigating SDC-5734.

Change-Id: Ibea1390c4dbcda65b0782392c475b5c9d9305c48
Reviewed-on: https://review.streamsets.net/7498
Reviewed-by: Virag Kothari <virag@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>