SDC-5816.  KUDU UPSERT and UPDATE cause null columns
When the Default Operation is set to UPSERT or UPDATE
the Kudu processor would update *all* columns in the Kudu
Table.  If there was a Kudu column which did not have
a corresponding field in the record - the Kudu column
would be set to null.

The example for this is:

The sample kudu table has 4 columns: id, fname, lname and address.
Insert the following JSON data:
{  "id": "1442aa76-22a0-49aa-8578-6edce9465171",
"fname": “Joe”,
"address": “12 Main St.” }

The data is inserted correctly:
id | fname | lname | address  1442aa76-22a0-49aa-8578-6edce9465171 | Joe | NULL | 12 Main St.

When we UPSERT this row with this JSON data:
{  "id": "1442aa76-22a0-49aa-8578-6edce9465171",
"lname": “Smith”  }
We see:
id | fname | lname | address  1442aa76-22a0-49aa-8578-6edce9465171 | NULL | Smith | NULL

Columns which were not specified in the incoming JSON are set to null.

This is fixed so that UPSERT and UPDATE will update all the fields which
are explicitly specified in the UI and will perform implicit mapping
fields in the record whose names match columns in the Kudu table are
also included in the fields to UPSERT or UPDATE.  Fields which are not
mapped in the UI - and do not have corresponding names between the
record and the Kudu columns will not be updated.

If this is not the desired behavior, and there are columns which
should not be updated - use a field renamer or field remover and
ensure that the field names no longer match the Kudu column names.

Change-Id: I4c5bf71855007c51525be64e253d98fe91292153
Reviewed-on: https://review.streamsets.net/7583
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
(cherry picked from commit ff61219216da363f56d7fefdf5aa01ddb77ccd67)
Reviewed-on: https://review.streamsets.net/7585
Reviewed-by: Bob Plotts <bob@streamsets.com>