SDC-5838: FullPipeBatch.skipBatch does not properly initialize empty event lane
While destroying multi-threaded pipeline, we destroy origin pipe and
them move on to destroy all pipe runners. Since the origin was already
destroyed and we need to perform one last run for remaining stages, we
always call FullStagePipe.skipStage for the origin stage.

This call is suppose to fill all internal structures of the
FullStagePipe as if the origin would actually run - just with empty or
default values. The code was correctly generating empty structure for
data lanes, but was missing doing the same for event lane.

Change-Id: I80fea76f00824b4d21d42febbf3915e40f2f9f86
Reviewed-on: https://review.streamsets.net/7696
Reviewed-by: Bob Plotts <bob@streamsets.com>
(cherry picked from commit 0c32b91155eadbdd2214e795b45ee58c048e6fb7)
Reviewed-on: https://review.streamsets.net/7731
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>