SDC-5874: Remove ElasticsearchConfig.PREFIX
It seems that we've introduced the prefix in believe that all stages
will use the same prefix, but that is not the case - destination uses
different one then the origin which leads to only problems. Let's drop
the static variable and let's force each method to properly pass the
desired prefix as we're already doing in other components (like hive).

Change-Id: I1ad51617fbb08cb7b6844bb6c4bcc29d0e07447c
Reviewed-on: https://review.streamsets.net/7725
Reviewed-by: Jeff Evans <jeff@streamsets.com>