SDC-5651: NPE in SyncPreviewer.stop()
The previewer is written in a way that it does not expects the
previewStatus to be null. But it can be if the previewer was just
created and never "used" from whatever reason. Rather then adding null
checks to all the places where we have this assumption, I've added a new
state - CREATED. Hence making sure that the variable will actually never
be null.

Change-Id: I52cfc07d7f52e4a11389655ab263dcf7cf95678e
Reviewed-on: https://review.streamsets.net/7729
Reviewed-by: Pat Patterson <pat@streamsets.com>
(cherry picked from commit 7f979d17b02a61c99fe79c62092bfded6c7139f5)
Reviewed-on: https://review.streamsets.net/7746
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>