SDC-5912. Remove netty exclusion to fix Cassandra tests
Also removed flaky call to cleanEmbeddedCassandra which simply drops
all keyspaces. We already clean these up in the @After method.

A race condition likely due to eventual consistency causes a failure
when dropping keyspaces sometimes since it tries to perform this
operation twice in a row at the end of the test suite.

Change-Id: Ia3f40d32b39c88e20c61901904c0d5125dc7588a
Reviewed-on: https://review.streamsets.net/7846
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>