SDC-5911: JDBC Producer sends empty batch to driver, resulting in exception on Teradata
It turns out that we already have a logic to not run empty statement
batches. There is however a small bug in the logic - if adding record to
the batch fails, we could issue empty batch.

This path moves th logic that adds statement to the "to be executed"
array only if adding at least one record succeeded.

Change-Id: I2916f7a465c6a44c85aa29a9bd7bc0c4f0f9dc0f
Reviewed-on: https://review.streamsets.net/7894
Reviewed-by: Pat Patterson <pat@streamsets.com>