SDC-6030: Bundles: Redact pipeline configuration files
This is pretty much rewrite of the pipeline generator. Instead of just
copying blindly files from data directory, this patch will use our
native stores to retrieve files that we're interested in. The benefit of
this approach is that we don't have to deal with various race conditions
when the files would be half written, since they are concurrently
edited. This also enabled us to do very simple redaction of password
fields as we have the pipeline in handy Java objects.

Change-Id: I47ac902ffb7334c82bd4e9a38adb09ac91774756
Reviewed-on: https://review.streamsets.net/8026
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>