SDC-4707.  Email Executor.
Note - tests will be writen for JIRA SDC-5976.

This executor will send an email to a specified list of
recipients when a condition is met.

The UI's Condition field supports an EL
which evaluates to true or false.

The Executor's UI permits several emails to be set
up in a specific instance of the Executor.

You could have an email to be sent
when there is "no-more-data" on the JDBC Query Origin.

When a Condition evaluates to true, the other
fields in the group - list of emails, subject
and body are processed in case they
contain any ELs and an email is sent to the
recipients in the list.

The email field can also contain ELs, in case the
desired email address is in a field in the
incoming record.

The recipient list is de-duped prior to
sending the email.

Change-Id: Id6babd6b58150704b7654259df73fd8738805be5
Reviewed-on: https://review.streamsets.net/8058
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Adam Kunicki <adam@streamsets.com>