SDC-6158. Exception on release class loader on stage destroy can cause failure to set the thread's context class loader.
It happens that we have a block where we do stageBean.releaseClassLoader and then reset the thread classloader to Container classloader from stage class loader.
But because of the exception java.lang.IllegalStateException: Object has already been returned to this pool or is invalid, releaseClassLoader failed and we never reset the classloader of the thread to Container class loader. This patch fixes that.

Change-Id: I3d3ff6846d374eda687ab2cd9c5c198f3a0cca98
Reviewed-on: https://review.streamsets.net/8228
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>