SDC-6196: Make properties on HDFS Target Output Files tab depend on File Type rather then Data Format
We have several config options available on "Output Files" tab that
depends on values in "Data Format" field (different tab). This means
that user have to click on Data Format tab, select file format, and then
return to Output Files tab to fill them up which is very confusing.

It seems that the only reason why we have this dependency is to hide a
bunch of configs that are not applicable to whole file format - but for
user to use whole file format on Data Format tab, they already have to
also select "file type" on the "output files tab." And we do have checks
that those two config options are equal. Hence to improve usability, I
suggest switching the dependency.

Change-Id: I56937a7ff3642bfa47724327a606fa63051adaba
Reviewed-on: https://review.streamsets.net/8272
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>