SDC-5754: Pipeline history should include application id
The application id is already in the history file, it's just not
displayed. Rather then doing some if-else conditions in UI, I took the
approach of updating the message itself with the application id. This
way people who needs to use the ID in a script can use the JSON and take
the id trivially from there, users using UI can see it.

Change-Id: Idc88f92c33aa4ef0bddaae6908aa57f838b98e21
Reviewed-on: https://review.streamsets.net/8389
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Bob Plotts <bob@streamsets.com>