SDC-6334: Split S3Config.java between origin and destination
Currently, the S3Config is shared between source and destination and
also contains bucket configuration. This means that bucket is a
constant. To support RecordEL in the bucket for the only destination, we
need to refactor the bean and split it between origin and destination.

This patch on high level does:

* Renames shared S3Config.java to S3ConnectionBaseConfig.java and moves
  it to shared package while making it an abstract class.
* Creates two subclasses - one for origin, second for destination.
* Moves the config "Bucket" to the subclasses.

Actually supporting RecordEL in bucket name for destination will be
covered by SDC-5490. I want to separate the JIRAs since otherwise it
would be too large of a change.

Change-Id: I53675bab878fad1ba60bed83fb13430a1c577637
Reviewed-on: https://review.streamsets.net/8590
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>
Reviewed-on: https://review.streamsets.net/9279
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>