SDC-6365. Hdfs/Local FS Target with whole file on exception renames the file to final name.
This patch fixes the case where in a whole file is not fully copied and an exception is thrown while copying, hdfs/local fs renames the file anyway (ending up with corrupted/0 byte files). This is because on IOException a stage exception is thrown by hdfs stage which makes hdfs destroy to be called which renames the files which are not closed (though this holds good for normal file formats, for whole file we should not close), for whole file we will close the file immediately after file is fully written. In case of not renaming the file, the file will be copied from scratch.

Change-Id: I40719020b147e0936398a0b090bff7df8a1e58a8
Reviewed-on: https://review.streamsets.net/8650
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>