SDC-6407: LDAP login needs a way to keep usernames lowercase
After some research it seems that this is actually fairly common
behavior in Hadoop land - to lowercase user names coming from LDAP.
Hence this patch adds similar functionality via a stage configuration
(the one stored in sdc.properties). I've decided to do this rather then
the pipeline configuration to prohibit user to change the value as it's
sort of an attack surface.

Change-Id: I93a8ac36c800dc0df575a30c9f1f4c66a39a8e73
Reviewed-on: https://review.streamsets.net/8789
Reviewed-by: Bob Plotts <bob@streamsets.com>