SDC-4677: Enhance HDFS Origin to read from non hdfs file systems
This check removes the validation that HDFS uri must start with hdfs://
and thus enables to use of any arbitrary file systems that have Hadoop
FS implementation (S3, ADL, ADSL, ...). We're not explicitly adding any
jars and such, so it's expected that users who will need to explore
this, will install them on their own.

Change-Id: Id56b4201c90db49e535abfe9cf4e409719b21132
Reviewed-on: https://review.streamsets.net/8791
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Bob Plotts <bob@streamsets.com>