SDC-6470. Generate a query for single row operation when 'Use Multi-Row Operation' is not selected
This is a regression since 2.6 release. Fixing DELETE operation.

When 'Use Multi-Row Operation' option is not selected, we should generate
a query for single-row operation, which is DELETE xxxx WHERE colName = value.
2.6 release generates DELETE xx WHERE colName IN (value) for both
single-row and multi-row operations, but this query is not supported in
MS SQL and DB2 (when table has multiple primary keys).
We should support at least single-row operation for all databases.

Tested on all databases and single-row op workd on all.

Change-Id: Ica37fed003c7acbb582af2f7b33837075c81849a
Reviewed-on: https://review.streamsets.net/8917
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>