SDC-6505.  Enhance events Directory Origin publishes
Add a no-more-data event to Directory origin,
and enhance the FINISHED_FILE event.

For the FINISHED_FILE event, we added the record count of good
records, and the count of error records.  The file name
was already published in the event.

Also added no-more-data events.  These are
generated by Directory Origin after the origin runs out
of data and the "Batch Wait Time" has elapsed.

This permits late-arriving files to be processed, and still gives
control over the delay for triggering the no-more-data-event.

One use case for no-more-data is to stop the pipeline:

By hooking the Directory Origin's event stream to
a Stream Selector and setting the Stream Selector to send
${record:eventType() == "no-more-data"}
events to the Pipeline Finisher Executor this will have
the effect of stopping the pipeline when there
is no more data and the Batch Wait Time has elapsed.

Three attributes are included in the no-more-data event:
record-count, error-count, and file-count.

record-count which is the number of records processed
since the pipeline started this run.  This number only
includes valid records which have been added to a batch.

error-count is the number of records which have been
sent to the error destination.

file-count which is the number of files processed
since the pipeline started this run.

None of the values are persisted - they are all
reset to 0 after the no-more-data event is sent out.

Change-Id: I6b5b6d585983d8dfd5c93b4791a265dad04efb05
Reviewed-on: https://review.streamsets.net/8916
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>