SDC-6519. Check for DynamoDB table existence before attempting deletion
This adds a user friendly error message if they requested
to reset the offset but we can't delete the table.

There is an unavoidable condition where if you kill SDC
after starting the pipeline but no offsets were committed
to the local offset.json but the table was created already.

In this case SDC thinks it needs to reset offset because
offset.json was not written to disk, however a dynamodb table
already exists with offsets.

Change-Id: I9a862561fcfec2d7177efd0478f0639d27dcf71e
Reviewed-on: https://review.streamsets.net/8943
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>